+++ src/main/java/com/telepathicgrunt/bumblezone/mixin/entities/LivingEntityTickMixin.java
@@ -1,21 +1,19 @@
 package com.telepathicgrunt.bumblezone.mixin.entities;
 
-import com.telepathicgrunt.bumblezone.entities.BeeAggression;
 import com.telepathicgrunt.bumblezone.entities.EntityTeleportationHookup;
-import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.entity.LivingEntity;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@Mixin(PlayerEntity.class)
-public class PlayerTickMixin {
+@Mixin(LivingEntity.class)
+public class LivingEntityTickMixin {
 
-    // Handles where wrath of the hive can be on,
-    // change player fog color when effect is active
+    // Handles teleportation
     @Inject(method = "tick",
             at = @At(value = "TAIL"))
-    private void thebumblezone_onPlayerTick(CallbackInfo ci) {
-        BeeAggression.playerTick((PlayerEntity) (Object) this);
+    private void thebumblezone_onLivingEntityTick(CallbackInfo ci) {
+        EntityTeleportationHookup.entityTick((LivingEntity) (Object) this);
     }
 }