+++ src/main/java/com/telepathicgrunt/bumblezone/mixin/entities/PlayerDimensionChangeMixin.java
@@ -1,19 +1,27 @@
 package com.telepathicgrunt.bumblezone.mixin.entities;
 
-import com.telepathicgrunt.bumblezone.entities.PlayerTeleportationBackend;
-import net.minecraft.server.network.ServerPlayerEntity;
+import com.telepathicgrunt.bumblezone.entities.EntityTeleportationBackend;
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.LivingEntity;
 import net.minecraft.server.world.ServerWorld;
+import net.minecraft.world.World;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-@Mixin(ServerPlayerEntity.class)
+@Mixin(Entity.class)
 public class PlayerDimensionChangeMixin {
-    // Handles storing of past non-bumblezone dimension the player is leaving
-    @Inject(method = "worldChanged(Lnet/minecraft/server/world/ServerWorld;)V",
+
+    @Shadow
+    public World world;
+
+    // Handles storing of past non-bumblezone dimension the entity is leaving
+    @Inject(method = "moveToWorld(Lnet/minecraft/server/world/ServerWorld;)Lnet/minecraft/entity/Entity;",
             at = @At(value = "HEAD"))
-    private void thebumblezone_onDimensionChange(ServerWorld origin, CallbackInfo ci) {
-        PlayerTeleportationBackend.playerLeavingBz(origin.getRegistryKey().getValue(), ((ServerPlayerEntity)(Object)this));
+    private void thebumblezone_onDimensionChange(ServerWorld destination, CallbackInfoReturnable<Entity> cir) {
+        EntityTeleportationBackend.playerLeavingBz(world.getRegistryKey().getValue(), ((Entity)(Object)this));
     }
 }